delete Game component
