+++ src/board/Board.test.jsx
@@ -1,4 +1,4 @@
-import { screen, render } from "@testing-library/react";
+import { screen, render, fireEvent } from "@testing-library/react";
 import Board from "./Board";
 
 test("has the correct number of rows", () => {
@@ -19,4 +19,59 @@ test("has the correct number of rows", () => {
   expect(screen.getByText("7")).toBeVisible();
   expect(screen.getByText("8")).toBeVisible();
   expect(screen.getByText("9")).toBeVisible();
+});
+
+describe("when user clicks a cell", () => {
+  test("board is updated with correct marker", () => {
+    const mockChangeTurn = jest.fn();
+    render(
+      <Board
+        setAlertVisible={jest.fn()}
+        changeTurn={mockChangeTurn}
+        currentPlayer="z"
+      />
+    );
+
+    const cellOne = screen.getByText("1");
+    fireEvent.click(cellOne);
+
+    expect(screen.queryByText("1")).not.toBeInTheDocument();
+    expect(cellOne).toHaveTextContent("z");
+  });
+
+  test("changes turn", () => {
+    const mockChangeTurn = jest.fn();
+    render(
+      <Board
+        setAlertVisible={jest.fn()}
+        changeTurn={mockChangeTurn}
+        currentPlayer="z"
+      />
+    );
+
+    const cellOne = screen.getByText("1");
+    fireEvent.click(cellOne);
+
+    expect(mockChangeTurn).toHaveBeenCalled();
+  });
+
+  test("other cells remain unchanged", () => {
+    const mockChangeTurn = jest.fn();
+    render(
+      <Board
+        setAlertVisible={jest.fn()}
+        changeTurn={mockChangeTurn}
+        currentPlayer="z"
+      />
+    );
+
+    const cellOne = screen.getByText("1");
+    const cellTwo = screen.getByText("2");
+    const cellThree = screen.getByText("3");
+    fireEvent.click(cellOne);
+
+    expect(cellOne).toHaveTextContent("z");
+    expect(cellTwo).toHaveTextContent("2");
+    expect(cellThree).toHaveTextContent("3");
+  });
 });
